Move binding steps to 'addBinding' folder
