+++ src/commands/addBinding/settingSteps/BooleanPromptStep.ts
@@ -7,16 +7,16 @@ import { QuickPickItem } from "vscode";
 import { AzureWizardPromptStep } from "vscode-azureextensionui";
 import { ext } from "../../../extensionVariables";
 import { IFunctionSetting } from "../../../templates/IFunctionSetting";
-import { IFunctionWizardContext } from "../IFunctionWizardContext";
+import { IBindingWizardContext } from "../IBindingWizardContext";
 
-export class BooleanPromptStep extends AzureWizardPromptStep<IFunctionWizardContext> {
+export class BooleanPromptStep extends AzureWizardPromptStep<IBindingWizardContext> {
     private readonly _setting: IFunctionSetting;
     constructor(setting: IFunctionSetting) {
         super();
         this._setting = setting;
     }
 
-    public async prompt(wizardContext: IFunctionWizardContext): Promise<void> {
+    public async prompt(wizardContext: IBindingWizardContext): Promise<void> {
         const picks: QuickPickItem[] = [
             { label: 'true', description: '' },
             { label: 'false', description: '' }
@@ -24,7 +24,7 @@ export class BooleanPromptStep extends AzureWizardPromptStep<IFunctionWizardCont
         wizardContext[this._setting.name] = (await ext.ui.showQuickPick(picks, { placeHolder: this._setting.label })).label;
     }
 
-    public shouldPrompt(wizardContext: IFunctionWizardContext): boolean {
+    public shouldPrompt(wizardContext: IBindingWizardContext): boolean {
         return !wizardContext[this._setting.name];
     }
 }
