+++ src/commands/addBinding/settingSteps/EnumPromptStep.ts
@@ -6,21 +6,21 @@
 import { AzureWizardPromptStep, IAzureQuickPickItem } from "vscode-azureextensionui";
 import { ext } from "../../../extensionVariables";
 import { IFunctionSetting } from "../../../templates/IFunctionSetting";
-import { IFunctionWizardContext } from "../IFunctionWizardContext";
+import { IBindingWizardContext } from "../IBindingWizardContext";
 
-export class EnumPromptStep extends AzureWizardPromptStep<IFunctionWizardContext> {
+export class EnumPromptStep extends AzureWizardPromptStep<IBindingWizardContext> {
     private readonly _setting: IFunctionSetting;
     constructor(setting: IFunctionSetting) {
         super();
         this._setting = setting;
     }
 
-    public async prompt(wizardContext: IFunctionWizardContext): Promise<void> {
+    public async prompt(wizardContext: IBindingWizardContext): Promise<void> {
         const picks: IAzureQuickPickItem<string>[] = this._setting.enums.map(e => { return { data: e.value, label: e.displayName }; });
         wizardContext[this._setting.name] = (await ext.ui.showQuickPick(picks, { placeHolder: this._setting.label })).data;
     }
 
-    public shouldPrompt(wizardContext: IFunctionWizardContext): boolean {
+    public shouldPrompt(wizardContext: IBindingWizardContext): boolean {
         return !wizardContext[this._setting.name];
     }
 }
