+++ src/commands/addBinding/settingSteps/LocalAppSettingCreateStep.ts
@@ -9,9 +9,9 @@ import { localSettingsFileName } from '../../../constants';
 import { setLocalAppSetting } from '../../../LocalAppSettings';
 import { localize } from '../../../localize';
 import { nonNullProp } from '../../../utils/nonNull';
-import { IFunctionWizardContext } from '../IFunctionWizardContext';
+import { IBindingWizardContext } from '../IBindingWizardContext';
 
-export class LocalAppSettingCreateStep extends AzureWizardExecuteStep<IFunctionWizardContext> {
+export class LocalAppSettingCreateStep extends AzureWizardExecuteStep<IBindingWizardContext> {
     public priority: number = 210;
 
     private readonly _nameKey: string;
@@ -23,13 +23,13 @@ export class LocalAppSettingCreateStep extends AzureWizardExecuteStep<IFunctionW
         this._valueKey = valueKey;
     }
 
-    public async execute(wizardContext: IFunctionWizardContext, progress: Progress<{ message?: string | undefined; increment?: number | undefined }>): Promise<void> {
+    public async execute(wizardContext: IBindingWizardContext, progress: Progress<{ message?: string | undefined; increment?: number | undefined }>): Promise<void> {
         progress.report({ message: localize('updatingLocalSettings', 'Updating {0}...', localSettingsFileName) });
         // tslint:disable-next-line: no-unsafe-any no-any
         await setLocalAppSetting(wizardContext.projectPath, nonNullProp(wizardContext, <any>this._nameKey), nonNullProp(wizardContext, <any>this._valueKey));
     }
 
-    public shouldExecute(wizardContext: IFunctionWizardContext): boolean {
+    public shouldExecute(wizardContext: IBindingWizardContext): boolean {
         return !!wizardContext[this._valueKey];
     }
 }
